chore: split QualificationDetailsController.cs into a new QualificationSearchController.cs
